% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaGeometricLikehood.R
\name{LogL.Dmu}
\alias{LogL.Dmu}
\title{Title}
\usage{
LogL.Dmu(ThetaBar.Par, M.Par)
}
\arguments{
\item{ThetaBar.Par}{Beta distribution mean paramter}

\item{M.Par}{Beta distribution precision paramter}
}
\description{
Title
}
