% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPMLEstimateFunction.R
\name{NPMLEstimateFunction}
\alias{NPMLEstimateFunction}
\title{Title}
\usage{
NPMLEstimateFunction(N, obs)
}
\arguments{
\item{N:}{the geometric sampling result}

\item{obs:}{the censored status of geometric sampling}
}
\value{
the discrete probability mass of the underlying distribution. UU is the probability mass, pp is the corresponding probability on that probability mass
}
\description{
Title
}
