% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAV.R
\name{PAV}
\alias{PAV}
\title{Eliminating the Unmeasured Confounders and Estimating Causal Effect.}
\usage{
PAV(formula, data = data, sdmethod = c("normal", "bootstrap", "all"),
  x1.name = "x1", x2.name = "x2", boots.no = NULL)
}
\arguments{
\item{formula}{an object of class "formula" : a symbolic description of the model to be fitted}

\item{data}{an optional data frame containing the variables in the model.}

\item{sdmethod}{Choosing a method for estimating variance}

\item{x1.name}{the name of exposure}

\item{x2.name}{the name of post-exposure}

\item{boots.no}{the number of bootstrap}
}
\value{
\code{formula} the model

\code{bootstrap} the casual effect coefficients when the variance is calculated by bootstrap

\code{normal} the casual effect coefficients when the variance is calculated in theory

\code{Quant} the confidence interval calculated by bootstrap
}
\description{
Eliminating the Unmeasured Confounders and Estimating Causal Effect.
}
\examples{
PAV <-  function(y~x1+x2, data = data, sdmethod ="normal",x1.name = "x1",x2.name = "x2",boots.no = NULL)




}
