% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledEstimateFunction.R
\name{PooledEstimateFunction}
\alias{PooledEstimateFunction}
\title{Title}
\usage{
PooledEstimateFunction(N, cen)
}
\arguments{
\item{N}{the geometric sampling result}

\item{cen}{the censored status of geometric samplings}
}
\description{
Title
}
