% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratEstimateFunction.R
\name{StratEstimateFunction}
\alias{StratEstimateFunction}
\title{Title}
\usage{
StratEstimateFunction(N, cen)
}
\arguments{
\item{N}{the geometric sampling result}

\item{cen}{the censored status of geometric sampling}
}
\description{
Title
}
