% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPMLEstimateFunction.R
\name{NPMLEstimateFunction}
\alias{NPMLEstimateFunction}
\title{Title}
\usage{
NPMLEstimateFunction(N, cen, d = c(1:length(N)))
}
\arguments{
\item{N:}{the geometric sampling result}

\item{cen:}{the censored status of geometric sampling}

\item{d:}{an index to select part of the data, default is 1:N. make it easy for the boot package.}
}
\value{
the discrete probability mass of the underlying distribution. UU is the probability mass, pp is the corresponding probability on that probability mass
}
\description{
Title
}
